requirements.txt
# Core dependencies for ShadowFS
fusepy>=3.0.1
pyyaml>=6.0
jinja2>=3.1.2

# Optional features (uncomment as needed)
# markdown>=3.4.0  # For markdown transform
# prometheus-client>=0.16.0  # For metrics export